test: separate endpoint matching coverage from sharing fixtures
