+++ test/routing_tests.cc
@@ -14,6 +14,8 @@
 
 namespace fs = std::filesystem;
 
+// Geometry tests allow endpoint matching penalties; cost_budget_test covers
+// tight limits.
 constexpr auto kTestMaxCost = osr::cost_t{100'000U};
 
 std::string extract_and_route(
@@ -47,9 +49,7 @@ TEST(routing, foot_island) {
   auto const to = osr::location{49.873023, 8.651523, osr::level_t{0.F}};
   EXPECT_EQ(
       R"({"type":"FeatureCollection","metadata":{},"features":[{"type":"Feature","properties":{"level":0E0,"osm_way_id":0,"cost":68,"distance":3},"geometry":{"type":"LineString","coordinates":[[8.6515174E0,4.98727447E1],[8.6515563E0,4.98727556E1]]}},{"type":"Feature","properties":{"level":0E0,"osm_way_id":22937760,"cost":2,"distance":2},"geometry":{"type":"LineString","coordinates":[[8.6515563E0,4.98727556E1],[8.6515406E0,4.98727706E1]]}},{"type":"Feature","properties":{"level":0E0,"osm_way_id":22937760,"cost":10,"distance":10},"geometry":{"type":"LineString","coordinates":[[8.6515406E0,4.98727706E1],[8.6514528E0,4.98728367E1]]}},{"type":"Feature","properties":{"level":0E0,"osm_way_id":847710844,"cost":6,"distance":7},"geometry":{"type":"LineString","coordinates":[[8.6514528E0,4.98728367E1],[8.6514754E0,4.98728959E1]]}},{"type":"Feature","properties":{"level":0E0,"osm_way_id":23511479,"cost":8,"distance":8},"geometry":{"type":"LineString","coordinates":[[8.6514754E0,4.98728959E1],[8.651539E0,4.98729497E1]]}},{"type":"Feature","properties":{"level":0E0,"osm_way_id":23511479,"cost":2,"distance":2},"geometry":{"type":"LineString","coordinates":[[8.651539E0,4.98729497E1],[8.6515596E0,4.98729618E1]]}},{"type":"Feature","properties":{"level":0E0,"osm_way_id":0,"cost":69,"distance":6},"geometry":{"type":"LineString","coordinates":[[8.6515596E0,4.98729618E1],[8.651493334251755E0,4.987300334757231E1]]}}]})",
-      extract_and_route("test/luisenplatz-darmstadt.osm.pbf", from, to,
-                        osr::foot<false, osr::elevator_tracking>::parameters{},
-                        osr::search_profile::kFoot));
+      extract_and_route("test/luisenplatz-darmstadt.osm.pbf", from, to));
 }
 
 TEST(routing, sharing_geofencing_does_not_restrict_walking) {
@@ -104,9 +104,7 @@ TEST(routing, foot_sidewalk_separate) {
   auto const to = osr::location{46.17249, 8.800507, osr::level_t{0.F}};
   EXPECT_EQ(
       R"({"type":"FeatureCollection","metadata":{},"features":[{"type":"Feature","properties":{"level":0E0,"osm_way_id":0,"cost":9,"distance":9},"geometry":{"type":"LineString","coordinates":[[8.799938095816453E0,4.617140034140192E1],[8.7999968E0,4.61714749E1]]}},{"type":"Feature","properties":{"level":0E0,"osm_way_id":313626022,"cost":52,"distance":52},"geometry":{"type":"LineString","coordinates":[[8.7999968E0,4.61714749E1],[8.8001826E0,4.61717214E1],[8.8002241E0,4.61717698E1],[8.8002659E0,4.61718363E1],[8.8002881E0,4.61718996E1]]}},{"type":"Feature","properties":{"level":0E0,"osm_way_id":37078706,"cost":157,"distance":22},"geometry":{"type":"LineString","coordinates":[[8.8002881E0,4.61718996E1],[8.800417E0,4.61718962E1],[8.8005511E0,4.61719431E1]]}},{"type":"Feature","properties":{"level":0E0,"osm_way_id":352238316,"cost":10,"distance":11},"geometry":{"type":"LineString","coordinates":[[8.8005511E0,4.61719431E1],[8.8005367E0,4.61720098E1],[8.8005388E0,4.61720441E1]]}},{"type":"Feature","properties":{"level":0E0,"osm_way_id":352238316,"cost":29,"distance":32},"geometry":{"type":"LineString","coordinates":[[8.8005388E0,4.61720441E1],[8.8005458E0,4.61721137E1],[8.8005774E0,4.61723293E1]]}},{"type":"Feature","properties":{"level":0E0,"osm_way_id":1214507783,"cost":9,"distance":11},"geometry":{"type":"LineString","coordinates":[[8.8005774E0,4.61723293E1],[8.8006558E0,4.61724109E1]]}},{"type":"Feature","properties":{"level":0E0,"osm_way_id":359235347,"cost":10,"distance":12},"geometry":{"type":"LineString","coordinates":[[8.8006558E0,4.61724109E1],[8.800673E0,4.61724166E1],[8.8007979E0,4.61724583E1]]}},{"type":"Feature","properties":{"level":0E0,"osm_way_id":359235347,"cost":4,"distance":5},"geometry":{"type":"LineString","coordinates":[[8.8007979E0,4.61724583E1],[8.8008536E0,4.61724769E1]]}},{"type":"Feature","properties":{"level":0E0,"osm_way_id":359235347,"cost":16,"distance":19},"geometry":{"type":"LineString","coordinates":[[8.8008536E0,4.61724769E1],[8.8010842E0,4.61725453E1]]}},{"type":"Feature","properties":{"level":0E0,"osm_way_id":359235347,"cost":2,"distance":3},"geometry":{"type":"LineString","coordinates":[[8.8010842E0,4.61725453E1],[8.8011136E0,4.61725564E1]]}},{"type":"Feature","properties":{"level":0E0,"osm_way_id":1214531287,"cost":3,"distance":3},"geometry":{"type":"LineString","coordinates":[[8.8011136E0,4.61725564E1],[8.8010967E0,4.61725825E1]]}},{"type":"Feature","properties":{"level":0E0,"osm_way_id":1214531287,"cost":5,"distance":5},"geometry":{"type":"LineString","coordinates":[[8.8010967E0,4.61725825E1],[8.8010707E0,4.61726227E1]]}},{"type":"Feature","properties":{"level":0E0,"osm_way_id":313768797,"cost":7,"distance":7},"geometry":{"type":"LineString","coordinates":[[8.8010707E0,4.61726227E1],[8.8009919E0,4.6172598E1]]}},{"type":"Feature","properties":{"level":0E0,"osm_way_id":313768797,"cost":39,"distance":39},"geometry":{"type":"LineString","coordinates":[[8.8009919E0,4.6172598E1],[8.8007065E0,4.61725085E1],[8.8005142E0,4.61725232E1]]}},{"type":"Feature","properties":{"level":0E0,"osm_way_id":0,"cost":64,"distance":0},"geometry":{"type":"LineString","coordinates":[[8.8005142E0,4.61725232E1],[8.8005142E0,4.61725232E1]]}}]})",
-      extract_and_route("test/locarno.osm.pbf", from, to,
-                        osr::foot<false, osr::elevator_tracking>::parameters{},
-                        osr::search_profile::kFoot));
+      extract_and_route("test/locarno.osm.pbf", from, to));
 }
 
 TEST(routing, foot_corridor) {
@@ -116,9 +114,7 @@ TEST(routing, foot_corridor) {
       osr::location{51.547004658329, -0.05694437499428773, osr::level_t{0.F}};
   EXPECT_EQ(
       R"({"type":"FeatureCollection","metadata":{},"features":[{"type":"Feature","properties":{"level":0E0,"osm_way_id":0,"cost":1,"distance":0},"geometry":{"type":"LineString","coordinates":[[-5.6227579350404025E-2,5.154663151628954E1],[-5.62413E-2,5.15466308E1]]}},{"type":"Feature","properties":{"level":0E0,"osm_way_id":1057782512,"cost":6,"distance":6},"geometry":{"type":"LineString","coordinates":[[-5.62413E-2,5.15466308E1],[-5.62459E-2,5.15466862E1]]}},{"type":"Feature","properties":{"level":0E0,"osm_way_id":881103497,"cost":3,"distance":4},"geometry":{"type":"LineString","coordinates":[[-5.62459E-2,5.15466862E1],[-5.62474E-2,5.15467056E1],[-5.62483E-2,5.15467228E1]]}},{"type":"Feature","properties":{"level":0E0,"osm_way_id":881103497,"cost":6,"distance":7},"geometry":{"type":"LineString","coordinates":[[-5.62483E-2,5.15467228E1],[-5.62524E-2,5.15467815E1]]}},{"type":"Feature","properties":{"level":0E0,"osm_way_id":1079378708,"cost":5,"distance":6},"geometry":{"type":"LineString","coordinates":[[-5.62524E-2,5.15467815E1],[-5.62531E-2,5.15468061E1],[-5.62538E-2,5.15468316E1]]}},{"type":"Feature","properties":{"level":0E0,"osm_way_id":1079378709,"cost":3,"distance":4},"geometry":{"type":"LineString","coordinates":[[-5.62538E-2,5.15468316E1],[-5.62546E-2,5.15468484E1],[-5.62621E-2,5.15468663E1]]}},{"type":"Feature","properties":{"level":0E0,"osm_way_id":881103498,"cost":3,"distance":4},"geometry":{"type":"LineString","coordinates":[[-5.62621E-2,5.15468663E1],[-5.62628E-2,5.15469052E1]]}},{"type":"Feature","properties":{"level":0E0,"osm_way_id":1063762150,"cost":1,"distance":1},"geometry":{"type":"LineString","coordinates":[[-5.62628E-2,5.15469052E1],[-5.6263E-2,5.15469163E1]]}},{"type":"Feature","properties":{"level":0E0,"osm_way_id":1063762149,"cost":2,"distance":3},"geometry":{"type":"LineString","coordinates":[[-5.6263E-2,5.15469163E1],[-5.62635E-2,5.1546946E1]]}},{"type":"Feature","properties":{"level":0E0,"osm_way_id":881103496,"cost":7,"distance":9},"geometry":{"type":"LineString","coordinates":[[-5.62635E-2,5.1546946E1],[-5.62657E-2,5.15470109E1],[-5.62908E-2,5.15470107E1]]}},{"type":"Feature","properties":{"level":0E0,"osm_way_id":0,"cost":42,"distance":47},"geometry":{"type":"LineString","coordinates":[[-5.62908E-2,5.15470107E1],[-5.62911E-2,5.1547031E1],[-5.66186E-2,5.15470287E1],[-5.69331E-2,5.15470277E1],[-5.694401949978242E-2,5.1547027764967275E1]]}}]})",
-      extract_and_route("test/london-corridor.osm.pbf", from, to,
-                        osr::foot<false, osr::elevator_tracking>::parameters{},
-                        osr::search_profile::kFoot));
+      extract_and_route("test/london-corridor.osm.pbf", from, to));
 }
 
 TEST(routing, foot_stop_area) {
@@ -128,9 +124,7 @@ TEST(routing, foot_stop_area) {
       osr::location{48.725480463902784, 2.2588322597458728, osr::level_t{0.F}};
   EXPECT_EQ(
       R"({"type":"FeatureCollection","metadata":{},"features":[{"type":"Feature","properties":{"level":0E0,"osm_way_id":0,"cost":12,"distance":8},"geometry":{"type":"LineString","coordinates":[[2.261201037460255E0,4.872532607195976E1],[2.2612721E0,4.87253867E1]]}},{"type":"Feature","properties":{"level":0E0,"osm_way_id":1290289280,"cost":3,"distance":3},"geometry":{"type":"LineString","coordinates":[[2.2612721E0,4.87253867E1],[2.2612448E0,4.87254022E1]]}},{"type":"Feature","properties":{"level":0E0,"osm_way_id":892871341,"cost":6,"distance":7},"geometry":{"type":"LineString","coordinates":[[2.2612448E0,4.87254022E1],[2.2611731E0,4.87254412E1]]}},{"type":"Feature","properties":{"level":0E0,"osm_way_id":1419149027,"cost":7,"distance":9},"geometry":{"type":"LineString","coordinates":[[2.2611731E0,4.87254412E1],[2.2610764E0,4.87254927E1]]}},{"type":"Feature","properties":{"level":0E0,"osm_way_id":1419149026,"cost":16,"distance":19},"geometry":{"type":"LineString","coordinates":[[2.2610764E0,4.87254927E1],[2.260814E0,4.87254962E1]]}},{"type":"Feature","properties":{"level":0E0,"osm_way_id":1359929257,"cost":20,"distance":24},"geometry":{"type":"LineString","coordinates":[[2.260814E0,4.87254962E1],[2.2605996E0,4.87253278E1]]}},{"type":"Feature","properties":{"level":0E0,"osm_way_id":1429030767,"cost":10,"distance":12},"geometry":{"type":"LineString","coordinates":[[2.2605996E0,4.87253278E1],[2.2604973E0,4.87252419E1]]}},{"type":"Feature","properties":{"level":0E0,"osm_way_id":1429030769,"cost":2,"distance":2},"geometry":{"type":"LineString","coordinates":[[2.2604973E0,4.87252419E1],[2.2604761E0,4.87252248E1]]}},{"type":"Feature","properties":{"level":0E0,"osm_way_id":1429030768,"cost":7,"distance":8},"geometry":{"type":"LineString","coordinates":[[2.2604761E0,4.87252248E1],[2.2604036E0,4.87251661E1]]}},{"type":"Feature","properties":{"level":0E0,"osm_way_id":1359929257,"cost":22,"distance":26},"geometry":{"type":"LineString","coordinates":[[2.2604036E0,4.87251661E1],[2.2602208E0,4.87250223E1],[2.2601787E0,4.87249893E1]]}},{"type":"Feature","properties":{"level":1E0,"osm_way_id":222040526,"cost":25,"distance":30},"geometry":{"type":"LineString","coordinates":[[2.2601787E0,4.87249893E1],[2.2600672E0,4.87250507E1],[2.2600272E0,4.87250721E1],[2.2598576E0,4.87251608E1]]}},{"type":"Feature","properties":{"level":1E0,"osm_way_id":222040526,"cost":11,"distance":13},"geometry":{"type":"LineString","coordinates":[[2.2598576E0,4.87251608E1],[2.2597231E0,4.87252337E1]]}},{"type":"Feature","properties":{"level":1E0,"osm_way_id":222040526,"cost":4,"distance":5},"geometry":{"type":"LineString","coordinates":[[2.2597231E0,4.87252337E1],[2.2596673E0,4.87252638E1]]}},{"type":"Feature","properties":{"level":1E0,"osm_way_id":222040526,"cost":93,"distance":112},"geometry":{"type":"LineString","coordinates":[[2.2596673E0,4.87252638E1],[2.2596292E0,4.8725284E1],[2.2584789E0,4.87258964E1]]}},{"type":"Feature","properties":{"level":1E0,"osm_way_id":222040526,"cost":13,"distance":16},"geometry":{"type":"LineString","coordinates":[[2.2584789E0,4.87258964E1],[2.2583131E0,4.87259901E1]]}},{"type":"Feature","properties":{"level":1E0,"osm_way_id":642263621,"cost":8,"distance":10},"geometry":{"type":"LineString","coordinates":[[2.2583131E0,4.87259901E1],[2.2582647E0,4.87259491E1],[2.2582257E0,4.8725916E1]]}},{"type":"Feature","properties":{"level":1E0,"osm_way_id":642578765,"cost":2,"distance":3},"geometry":{"type":"LineString","coordinates":[[2.2582257E0,4.8725916E1],[2.2581937E0,4.87259329E1]]}},{"type":"Feature","properties":{"level":1E0,"osm_way_id":642263625,"cost":6,"distance":7},"geometry":{"type":"LineString","coordinates":[[2.2581937E0,4.87259329E1],[2.2581458E0,4.87258947E1],[2.2581625E0,4.87258856E1]]}},{"type":"Feature","properties":{"level":1E0,"osm_way_id":642578757,"cost":7,"distance":8},"geometry":{"type":"LineString","coordinates":[[2.2581625E0,4.87258856E1],[2.2582443E0,4.87258411E1]]}},{"type":"Feature","properties":{"level":5E-1,"osm_way_id":642263631,"cost":12,"distance":14},"geometry":{"type":"LineString","coordinates":[[2.2582443E0,4.87258411E1],[2.2581285E0,4.87257422E1]]}},{"type":"Feature","properties":{"level":0E0,"osm_way_id":1192127892,"cost":1,"distance":1},"geometry":{"type":"LineString","coordinates":[[2.2581285E0,4.87257422E1],[2.2581167E0,4.87257321E1]]}},{"type":"Feature","properties":{"level":0E0,"osm_way_id":1192127892,"cost":2,"distance":2},"geometry":{"type":"LineString","coordinates":[[2.2581167E0,4.87257321E1],[2.2581393E0,4.87257206E1]]}},{"type":"Feature","properties":{"level":0E0,"osm_way_id":0,"cost":99,"distance":10},"geometry":{"type":"LineString","coordinates":[[2.2581393E0,4.87257206E1],[2.2582250818254357E0,4.872579309702946E1]]}}]})",
-      extract_and_route("test/station-border.osm.pbf", from, to,
-                        osr::foot<false, osr::elevator_tracking>::parameters{},
-                        osr::search_profile::kFoot));
+      extract_and_route("test/station-border.osm.pbf", from, to));
 }
 
 TEST(routing, bus_private_gate) {
